+++ src/ucan/wnfs.ts
@@ -1,0 +1,8 @@
+export const PREFIX = "wnfs"
+
+export type Capability
+  = "CREATE"
+  | "REVISE"
+  | "SOFT_DELETE"
+  | "OVERWRITE"
+  | "SUPER_USER"